Move `memeid4s#Node` to its own package
