+++ memeid4s-cats/src/main/scala/memeid4s/cats/syntax.scala
@@ -22,7 +22,8 @@ import cats.effect.{Clock, Sync}
 import cats.syntax.functor._
 
 import memeid4s.digest.Digestible
-import memeid4s.scala.{Node, UUID}
+import memeid4s.node.Node
+import memeid4s.scala.UUID
 import memeid4s.time.{Posix, Time}
 
 trait syntax {
