+++ memeid4s/src/test/scala/memeid4s/scala/NodeSpec.scala
@@ -16,6 +16,7 @@
 
 package memeid4s.scala
 
+import memeid4s.node.Node
 import org.specs2.ScalaCheck
 import org.specs2.mutable.Specification
